pyproject.toml
[project]
name = "vma"
version = "0.1.1"
description = "Vulnerability Management Application"
authors = [
    {name = "ddeuterio",email = "daniel.garcia.anes@gmail.com"}
]
license = {text = "Apache-2.0 with Commons Clause"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "argparse (>=1.4.0,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "datetime (>=5.5,<6.0)",
    "requests (>=2.32.4,<3.0.0)",
    "ratelimit (>=2.2.1,<3.0.0)",
    "fastapi[response] (>=0.122.0,<0.123.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-asyncio (>=0.23.0,<0.24.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
    "google-cloud-storage (>=3.7.0,<4.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aiolimiter (>=1.1.0,<2.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)"
]

[project.scripts]
vma = "vma.app:main"

[tool.poetry]
packages = [
    { include = "vma", from = "src" }
]
include = [
    "src/vma/web/static/**/*",
    "src/vma/web/templates/**/*"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_mode = "auto"
